Update apt.conf for separate /usr/ , /boot/ and /tmp/ partitions
APT is able to remount some partitions when you want to install packages.
In this way you can have the /usr/ partition mounted in the read-only
mode, and APT can mount it with write permissions just for the
installation process. The same thing can be done when you use a separate
partition for /boot/ , i.e. you use external device for the /boot/
partition and you mount it by default in the read-only mode to avoid
accidental file system corruption. There's also the /tmp/ partition,
which is usually mounted with the "noexec" option, and this parameter
can cause some problems during the installation process, for instance,
the following errors can be observed:
  Can't exec "/tmp/*": Permission denied at
  /usr/share/perl/5.26/IPC/Open3.pm line 178.
  open2: exec of /tmp/* configure 1.179 failed: Permission denied at
  /usr/share/perl5/Debconf/ConfModule.pm line 59.